Add flags to karaf script so they don't interfere with JAVA_OPTS
Putting JAVA_OPTS between quotes in the karaf script led to issues with
the original parameters, but was needed because the parameters for the
crashReport script include spaces.
To fix this, don't touch the original JAVA_OPTS and write the flags for
the crashReport script directly into the karaf script (correctly
escaping spaces).