Preliminary integration of GitOps in Docker dynamic image
This commit adds to the dynamic Docker agent image a Git client
and the commands to initialize the Git repository in
/opt/instana/agent/etc. The remote repository and branch to be used
are provided via the INSTANA_GIT_REMOTE_REPOSITORY and
INSTANA_GIT_REMOTE_BRANCH environment variables, for example:

--env="INSTANA_GIT_REMOTE_REPOSITORY=https://github.com/mmanciop/gitops-test.git"
--env="INSTANA_GIT_REMOTE_BRANCH=master"

This change will not be merged to the main codebase. Instead, we are
going to extend the capability of the agent later on to perform the
required initialization steps automatically when the environment
variables listed above are discovered.