Expose Git repo init environment variables
If on bootstrap, the agent does not find a Git repository in
the <agent-installation-dir>/etc, but it has the environment
variables INSTANA_GIT_REMOTE_REPOSITORY and
INSTANA_GIT_REMOTE_BRANCH, the agent will create the Git
repository and configure one remote, with name configuration,
the value of INSTANA_GIT_REMOTE_REPOSITORY as the uri, and as
fetch refspec:

+refs/heads/<remoteBranchName>:refs/remotes/<remoteName>/<remoteBranchName>

This commit adds support for configuring the repository with
the following properties, which will be used by the agent to
determined where to fetch from:

[instana]
remotename = <which remote name to use, if multiple remotes present>
remoteuri = <pick which remote to use based on the URI, alternative to remotename>
remotebranch = <which remote branch to fetch>

This change also gets rid of the tracked branch concept, which
is not enough familiar to DevOps people. Instead, the agent
will use the following combinations for selecting the remote:

If the '[instana] remotename' property is set, find a remote
called like that property's value and fail if none is found.

If the '[instana] remoteuri' property is set, find the first
remote with that URI listed (could be more than one), and fail
if none is found.

Otherwise, use the first remote, and fails if none is found.
The branch to be fetched is selected as follows:

If the '[instana] remotebranch' property is set, use that
Otherwise, use FETCH_HEAD