Enable remote initialization of GitOps feature
Remotely initializing the host agent inside to use the GitOps feature
causes its restart.

Since we currently start the agent with `karaf daemon`, the ENTRYPOINT
of the container terminates when the agent restarts, causing the container
to be churned. Changing the mode to `karaf server` instead, the
/opt/instana/agent/bin/karaf will restart the stopping agent, enabling the
remote initialization of the GitOps feature.