update links to documentation
This was mainly motivated because this is one of the few repositories
that still linked to docs.instana.com (which has been a redirect to
instana.com/docs for a couple of years now).

While I was at it, I also replaced the other remaining instana.com
links with the appropriate ibm.com links (because instana.com is
defunct now as well).